Eeshlo fixed the iYFexport path problem and added a file selector button for it. Now win32 users should have less problems.
I added export of the "no shadow" flag (traceable) And changed the reflection
export scheme so now blender's RayMir is used as yafray's min_refle and the
amount of reflection is controled by the mir color. This seems more intuitive.
And now IOR is exported also for reflection only, wich is needed for fresnel.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@1843 954f8c5b-7b00-dc11-b283-0030488c597c