* Preferences for displaying the grid floor, X axis, Y axis, Z axis. (ported from tuhopuu2)
Controls are found in the 3D View Properties panel (with
the rest of the grid settings). This is more flexible for
people like me who hate the grid getting in the way when
modelling - turning off the grid floor and turning on the Z
axis gives a setup like in Wings 3D or Clay.

* Cleaned up the 3D View Properties and Background Image panel's layout while I was adding the new buttons (including change over to UiDefButBit).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@1994 954f8c5b-7b00-dc11-b283-0030488c597c