bugfix #1119
When using MotionBlur, the rendered coordinates were not correctly
adjusted, causing visual artifacts using the new AO, but it was also
visible for normal ray-shadow, which didn't become anti-aliased in
Mblur render.

Solved it with 2 globals now, to pass on blur offset to renderloop.
Leave it that way... the entire method used to jitter/AA/blur stuff
needs revision.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@2246 954f8c5b-7b00-dc11-b283-0030488c597c