This commit removes the glue from Blender, and with it the directories decrypt, deflate, encrypt, inflate, readstreamglue, sign, writeblenfile and writestreamglue.
Sirdude was so kind to modify the makefiles, so SCons and
Make are ready to build with the new Blender.

Visual Studio workspaces, solutions and projectfiles still need
to be updated (I'll do the .vcprojs and .sln myself after this commit).

Runtimes created with the Blender Publisher are not anymore
recognised - if you want these available, you'll have to convert
them first to .blends with the Publisher.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@2333 954f8c5b-7b00-dc11-b283-0030488c597c