Removed some testcode from yafray_Render.cpp Added the missing anti-aliasing pixel filter size and threshold parameters for manual AA control (disable 'Auto AA' button). Added support for yafray raytraced depth-of-field. Added extra panel for Camera in edit window to edit dof paramaters. The actual focus point will be drawn as a cross when camera 'ShowLimits' is enabled, similar to the aqsis code in tuhopuu.
Note to users: raytraced DoF is very slow, for best results, the default AA parameters are not good enough, especially with higher aperture values (more blur).
So for best results, disable 'Auto AA' and set the AA parameters yourself.
It works best with multi-pass AA ('AA passes' > 1) and a reasonable 'AA samples' value, something in the range 8 - 25 or even higher.
Currently the post-process DoF is not available in yafray, alternatives are being worked on.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@2792 954f8c5b-7b00-dc11-b283-0030488c597c