Recode of 'zbuffer selection'. I first was too afraid using the backbuffer selection method, since there's too much trash in drawobject.c and displists... But, I couldn't stand the zbuf selection to miss edges too often. Right now both methods are still in the code, and can be reviewed.
Please note for test:
- borderselect and circle-select still use old method
- new selection works directly on faces & edges
- subsurf optimal draw and selection works fine now

To detect the closest vertex or edge, i use the old 'spiral selection' code,
which is extremely accurate. If your mouse is close to an edge, it should
find it.

Most evidently I need to test speed... also, the disadvantage of backbuf
selection is the limited depth, which can be 16 bits (65 k faces) on systems.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3141 954f8c5b-7b00-dc11-b283-0030488c597c