Outliner features:
- Rename!
CTRL+leftmouse click on name, makes it a text button. Works for all items as
currently being displayed.
Most work was doing the Bones, which is a nightmare :) But it uses same
code as buttons in Armature-Editmode now, without even needing EditMode :)
When renaming a bone, the Outliner makes the Object active though.

- PageUp / PageDown keys
Do what you expect.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3223 954f8c5b-7b00-dc11-b283-0030488c597c