Ton broke a couple of things in his last commit including PET in rotation mode and local axis constraints on objects. Bringing that back and enabling PET in trackball rotate.
Changed the rotation manipulator drawing code to really align the Trackball rotate ball with the view (using getViewVector) so that it always looks centered on the selection.
This was particularly ugly in perspective mode with a selection far from the center of the screen: http://www.clubinfo.bdeb.qc.ca/~theeth/screenie.jpg

Moved getViewVector from transform_constraints.c to transform_generics.c since it is not really a constraint related function. Also made it independant on the TransInfo structure so it might be useful elsewhere too.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3984 954f8c5b-7b00-dc11-b283-0030488c597c