Shrink/Fatten behavior change. This time for good I hope.
Technically, it now works by getting the mouse motion in 3D (just like Translation/Grab), projecting it on the vertical view axis, using the vector length as the shrink/fatten factor. If the motion is downward (on the screen), shrink, if it is upward, fatten.
In layman terms: move up to fatten, down to shrink and it adapts to the viewport zoom, like Translation.

I changed to snapgrid factors to match those of Translation to, so it really acts like grabbing.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4005 954f8c5b-7b00-dc11-b283-0030488c597c