Since Cygwin isn't compatible with all of this fancy AVI codec business we need to do a check to be sure that we're not attempting to compile AVI codec related features on Cygwin.
Prior to Rob's commit today, this setting wasn't exposed to Cygwin by
default because it was hidden by a WITH_QUICKTIME ifdef but when Rob
cleaned up today he removed that and the WIN32 ifdef inspired Cygwin and
the free tools to try and compile these pieces. This resulted in an
undefined reference to get_avicodec_settings.

Rob: Please review this change to be sure its compatible with the rest of
your integration of the avicodec and quicktime. It seems to go flawlessly
here but without your OK I'm not comfortable making this permenant. I based
this modifcation on other instances of this check for FREE_WINDOWS so it
should be just fine.

As always if it breaks something blame Ton. :P


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4041 954f8c5b-7b00-dc11-b283-0030488c597c