Commented out some old cruft in the Makefile which seemed to be unrelated to the Cygwin platform. This produced some silly error message about files which haven't existed in the CVS repository for many months now. I can remember seeing this message for at least a year and since it has *never* changed I decided to investigate. Looks like unmaintained part of the build system that nobody pays attention to but just in case I've left it in tact with the following comment:
This code below seems outdated and doesn't even seem to apply to the
Windows platform. I've commented it out until someone can take a look at it
and confirm or deny. If it breaks anything blame Chris Burt / Desoto

Hopefully that's clear enough, but I really think this is a trivial change
and if after review nobody minds I'll simply erase the offending code.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4042 954f8c5b-7b00-dc11-b283-0030488c597c